+++ AutoMLOps/AutoMLOps.py
@@ -312,3 +312,9 @@ def pipeline(func: Optional[Callable] = None,
             func=func,
             name=name,
             description=description)
+
+def clear_cache():
+    """Deletes all temporary files stored in the cache directory."""
+    execute_process(f'rm -rf {OUTPUT_DIR}', to_null=False)    
+    make_dirs([OUTPUT_DIR])
+    logging.info('Cache cleared.')